refactor: 공통 사용 예외 메시지 Enum 분리
- 추후 운영 시간 예외 메시지도 포맷팅할 가능성을 고려하여 FormattedErrorMessage로 분리했습니다.
