+++ src/main/java/attendance/domain/Attendance.java
@@ -1,9 +1,9 @@
-package attendance;
+package attendance.domain;
+
+import attendance.util.FormattedErrorMessage;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
-import java.time.format.TextStyle;
-import java.util.Locale;
 
 public class Attendance {
 
@@ -17,9 +17,7 @@ public class Attendance {
     private void validateAttendDate(LocalDate attendDate) {
         DayOfWeek dayOfWeekOfAttendDate = attendDate.getDayOfWeek();
         if (dayOfWeekOfAttendDate == DayOfWeek.SATURDAY || dayOfWeekOfAttendDate == DayOfWeek.SUNDAY) {
-            throw new IllegalArgumentException("[ERROR] %d월 %d일 %s은(는) 등교일이 아닙니다"
-                    .formatted(attendDate.getMonthValue(), attendDate.getDayOfMonth(),
-                            attendDate.getDayOfWeek().getDisplayName(TextStyle.FULL, Locale.KOREAN)));
+            throw new IllegalArgumentException(FormattedErrorMessage.INVALID_ATTENDANCE_ERROR.getDateFormatMessage(attendDate));
         }
     }
 }
